+++ src/main/java/com/company/carrental/repository/CarTypeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarTypeRepository extends JpaRepository<CarType, Long> {
+public interface CarTypeRepository extends JpaRepository<CarType, Integer> {
     CarType findByVehicleType(VechicleType vehicleType);
 }
