+++ src/main/java/com/company/carrental/repository/ReservationRepository.java
@@ -9,9 +9,9 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
     List<Reservation> findByUser(User user);
 
     List<Reservation> findByCarCarIdAndStartDateLessThanEqualAndEndDateGreaterThanEqual(
-            Long carId, LocalDateTime endDate, LocalDateTime startDate);
+            Integer carId, LocalDateTime endDate, LocalDateTime startDate);
 }
